[HIB-91] Use pytest fixtures in scheduler tests
