+++ lambdas/scheduler/tests/test_dispatch.py
@@ -1,0 +1,3 @@
+
+def test_dispatch():
+    pass
