+++ lambdas/scheduler/tests/test_events.py
@@ -21,13 +21,3 @@ def test_make_new_granule_events(granule_events):
     assert all([
         isinstance(e, hyp3_events.NewGranuleEvent) for e in new_granule_events
     ])
-
-
-def test_events():
-    packages = utils.load_email_packages()
-
-    new_events = events.make_from(packages)
-
-    assert isinstance(new_events, list)
-    for event in new_events:
-        assert isinstance(event, hyp3_events.Hyp3Event)
